commit code

+++ src/resources/views/permissions/list.blade.php
@@ -4,9 +4,11 @@
             <h2 class="font-semibold text-xl text-gray-800 leading-tight">
                 {{ __('Permissions ') }}
             </h2>
-            <a href="{{ route('permissions.create') }}" class="bg-slate-700 text-sm rounded-md text-white px-3 py-2">
-                Create
-            </a>
+            @can('create permissions')
+                <a href="{{ route('permissions.create') }}" class="bg-slate-700 text-sm rounded-md text-white px-3 py-2">
+                    Create
+                </a>
+            @endcan
         </div>
     </x-slot>
 
@@ -30,9 +32,11 @@
                                 <td class="px-6 py-3 text-left">{{ $permission->name }}</td>
                                 <td class="px-6 py-3 text-left">{{ $permission->created_at->format('d M, Y') }}</td>
                                 <td class="px-6 py-3 text-center">
-                                    <a href="{{ route('permissions.edit', $permission->id) }}" class="bg-slate-700 hover:bg-slate-600 text-sm rounded-md text-white px-3 py-2 me-1">
-                                        Edit
-                                    </a>
+                                    @can('edit permissions')
+                                        <a href="{{ route('permissions.edit', $permission->id) }}" class="bg-slate-700 hover:bg-slate-600 text-sm rounded-md text-white px-3 py-2 me-1">
+                                            Edit
+                                        </a>
+                                    @endcan
                                     {{-- <form action="{{ route('permissions.destroy', $permission->id) }}" method="POST" style="display:inline-block;" onsubmit="return confirm('Are you sure?');">
                                         @csrf
                                         @method('DELETE')
@@ -40,9 +44,11 @@
                                             Delete
                                         </button>
                                     </form> --}}
-                                    <a href="javascript:void(0);" onclick="deletePermission({{ $permission->id }})" class="bg-red-600 hover:bg-red-500 text-sm rounded-md text-white px-3 py-2">
-                                        Delete
-                                    </a>
+                                    @can('delete permissions')
+                                        <a href="javascript:void(0);" onclick="deletePermission({{ $permission->id }})" class="bg-red-600 hover:bg-red-500 text-sm rounded-md text-white px-3 py-2">
+                                            Delete
+                                        </a>
+                                    @endcan
                                 </td>
                             </tr>
                         @endforeach
