+++ src/resources/views/roles/list.blade.php
@@ -4,9 +4,11 @@
             <h2 class="font-semibold text-xl text-gray-800 leading-tight">
                 {{ __('Roles ') }}
             </h2>
-            <a href="{{ route('roles.create') }}" class="bg-slate-700 text-sm rounded-md text-white px-3 py-2">
-                Create
-            </a>
+            @can('create roles')
+                <a href="{{ route('roles.create') }}" class="bg-slate-700 text-sm rounded-md text-white px-3 py-2">
+                    Create
+                </a>
+            @endcan
         </div>
     </x-slot>
 
@@ -32,12 +34,16 @@
                                 <td class="px-6 py-3 text-left">{{ $role->permissions->pluck('name')->implode(', ') }}</td>
                                 <td class="px-6 py-3 text-left">{{ $role->created_at->format('d M, Y') }}</td>
                                 <td class="px-6 py-3 text-center">
-                                    <a href="{{ route('roles.edit', $role->id) }}" class="bg-slate-700 hover:bg-slate-600 text-sm rounded-md text-white px-3 py-2 me-1">
-                                        Edit
-                                    </a>
-                                    <a href="javascript:void(0);" onclick="deleteRole({{ $role->id }})" class="bg-red-600 hover:bg-red-500 text-sm rounded-md text-white px-3 py-2">
-                                        Delete
-                                    </a>
+                                    @can('edit roles')
+                                        <a href="{{ route('roles.edit', $role->id) }}" class="bg-slate-700 hover:bg-slate-600 text-sm rounded-md text-white px-3 py-2 me-1">
+                                            Edit
+                                        </a>
+                                    @endcan
+                                    @can('delete roles')
+                                        <a href="javascript:void(0);" onclick="deleteRole({{ $role->id }})" class="bg-red-600 hover:bg-red-500 text-sm rounded-md text-white px-3 py-2">
+                                            Delete
+                                        </a>
+                                    @endcan
                                 </td>
                             </tr>
                         @endforeach
